% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_fator_capacidade.R
\name{calcula_fator_capacidade}
\alias{calcula_fator_capacidade}
\title{Calcula o fator de capacidade.}
\usage{
calcula_fator_capacidade(inflacao)
}
\arguments{
\item{inflacao}{Taxa anual de inflação de longo prazo no formato percentual. Valor numérco.}
}
\value{
Valor do fator de capacidade correspondente à taxa de inflação informada.
}
\description{
Esta função calcula o 'fator de capacidade' a partir da taxa de inflação
anual fornecida à função.
}
\examples{
calcula_fator_capacidade(4.5)
}
