% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_idade.R
\name{calcula_idade}
\alias{calcula_idade}
\title{Calcula a quantidade de anos entre duas datas}
\usage{
calcula_idade(dt_ini, dt_fim)
}
\arguments{
\item{dt_ini}{Data inicial. Pode ser uma string no formato "dd/mm/yyyy" ou 
um vetor que represente data (objeto da classe `Date` ou `POSIX`)}

\item{dt_fim}{Data final. Pode ser uma string no formato "dd/mm/yyyy" ou
um vetor que represente data (objeto da classe `Date` ou `POSIX`)}
}
\value{
Um vetor com a quantidade de anos entre as datas fornecidas.
}
\description{
A função calcula a quantidade de anos ente duas datas.
A data mais antiga deve ser fornecida como primeiro
argumento e a data mais recente como o segundo argumento.
}
