% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_multidecremento.R
\name{calcula_multidecremento}
\alias{calcula_multidecremento}
\title{Calcula probabilidades multidecrementais}
\usage{
calcula_multidecremento(
  qx = NULL,
  ix = NULL,
  tx = NULL,
  rx = NULL,
  retorno = c("qxT", "pxT")
)
}
\arguments{
\item{qx}{Vetor númerico contendo as probabilidades de morte. Argumento obrigatório.}

\item{ix}{Vetor numérico contendo as probabilidades de entrada em invalidez. Argumento obrigatório.}

\item{tx}{Vetor numérico contendo as probabilidades de saída do plano de benefícios (rotatividade/turnover).}

\item{rx}{Vetor numérico contendo as probabilidades de aposentadoria antecipada (retirement).}

\item{retorno}{String informando se devem ser retornadas as probabilidades multidecrementais de morte (`pxT`)
ou de sobrevivência (`pxT`)}
}
\value{
Um vetor númerico contendo as probabilidades multidecrementais.
}
\description{
Esta função calcula vetores de probabilidade considerando
dois ou mais decrementos.
}
\details{
Devem ser informados, no mínimo, os decrementos de morte e invalidez. 
Se forem declarados três decrementos estes serão considerados morte, invalidez e rotatividade.
}
