% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_Sproj.R
\name{calcula_Sproj}
\alias{calcula_Sproj}
\title{Calcula o salário projetado}
\usage{
calcula_Sproj(pxT, Sx, cs, t)
}
\arguments{
\item{pxT}{Probabilidades de sobrevivência multidecremental. Vetor numérico.}

\item{Sx}{Salário de contribuição mensal do servidor na data focal da avaliação atuarial.}

\item{cs}{Taxa real de crescimento anual salarial. Vetor numérico.}

\item{t}{Tempo de diferimento da projeção (em anos).}
}
\value{
Salários projetados para o período t
}
\description{
Função para realizar a projeção dos salários de contribuição
}
\details{
Os valores de `pxT` passados à função são obtidos com a função
[calcula_multidecremento()]

Esta função é usada internamente pela função `vasf()`
}
