% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_taxa_conjugada.R
\name{calcula_taxa_conjugada}
\alias{calcula_taxa_conjugada}
\title{Calcula taxa conjugada}
\usage{
calcula_taxa_conjugada(juros, incremento)
}
\arguments{
\item{juros}{Taxa de juros. Valor numérico em formato percentual.}

\item{incremento}{Taxa de crescimento dos salários ou benefícios. Valor numérico em formato percentual.}
}
\value{
Valor da taxa conjugada.
}
\description{
Calcula taxa conjugada
}
\examples{
calcula_taxa_conjugada(juros=4.5, incremento=2.5)

}
