% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_ien.R
\name{cn_ien}
\alias{cn_ien}
\title{Calcula a alíquota de contribuição normal anual}
\usage{
cn_ien(
  sx,
  inflacao,
  cs,
  sexo,
  idade_y,
  idade_r,
  idade_x,
  tmgm_act,
  tmgf_act,
  tmdm_ltb,
  tmdf_ltb,
  tmdm_act,
  tmdf_act,
  tmdm_txcjg_act,
  tmdf_txcjg_act,
  peso = c("sx", "axn_xr", "vasf_x1", "vasf_xr")
)
}
\arguments{
\item{sx}{salário mensal na data focal (x) da avaliação atuarial}

\item{inflacao}{valor da taxa de inflação de longo prazo em percentual}

\item{cs}{valor da taxa real de crescimento salarial em valor percentual}

\item{sexo}{sexo do servidor indicado no formato "M" e "F".}

\item{idade_y}{idade em anos na data de ingresso do servidor no sistema previdenciário (y)}

\item{idade_r}{idade em anos na data prevista de aposentadoria (r)}

\item{idade_x}{idade em anos na data focal (x)}

\item{tmgm_act}{tabua de comutação unidecremental masculina}

\item{tmgf_act}{tabua de comutação unidecremental feminina}

\item{tmdm_ltb}{tabua de mortalidade multidecremental masculina  - taxa de juros}

\item{tmdf_ltb}{tabua de mortalidade multidecremental feminina   - taxa de juros}

\item{tmdm_act}{tabua de comutação masculina multidecremental - taxa de juros}

\item{tmdf_act}{tabua de comutação feminina multidecremental - taxa de juros}

\item{tmdm_txcjg_act}{tabua de comutação masculina multidecremental - taxa conjugada}

\item{tmdf_txcjg_act}{tabua de comutação feminina multidecremental - taxa conjugada}

\item{peso}{peso de ponderação das alíquotas individuais, podendo assumir um dos seguintes
valores: `sx`, `axn_xr`, `vasf_x1` ou `vasf_xr`}

\item{dt_ini}{Data inicial. Pode ser uma string no formato "dd/mm/yyyy" ou 
um vetor que represente data (objeto da classe `Date` ou `POSIX`)}
}
\value{
Valor numérico indicando a alíquota de contribuição anual
}
\description{
A função calcula a alíquota de contribuição normal anual 
utilizando o método de financiamento atuarial Idade de
Entrada Normal, conforme definido no Art. 22 do Anexo VI
da Portaria n. 1467/2022 do Ministério da Previdência Social - MPS
}
