% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_comutacao.R
\name{cria_comutacao}
\alias{cria_comutacao}
\title{Cria tábua de comutação}
\usage{
cria_comutacao(tabua, juros)
}
\arguments{
\item{tabua}{Um objeto da classe `lifetable` do pacote `lifecontingencies` 
criado com a função `cria_tabua()`}

\item{juros}{Taxa de juros. Valor numérico em formato percentual.}
}
\value{
Um objeto da classe `actuarialtable` do pacote `lifecontingencies`
}
\description{
A função cria um objeto da classe `actuarialtable` do pacote
`lifecontingencies`
}
