% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_tabua.R
\name{cria_tabua}
\alias{cria_tabua}
\title{Cria tábua biométrica}
\usage{
cria_tabua(qx, md = FALSE)
}
\arguments{
\item{qx}{String contendo o nome da tábua a ser utilizada
obtidas com a função [calcula_multidecremento()] ou um vetor
numérico contendo probabilidades multidecrementais, caso
em que o argumento `md` (multidecremento) deve ser `TRUE`}

\item{md}{Indica se o argumento `qx` está recebendo um
vetor de probabilidades multidecrementalis ou não.
Se `FALSE` (opção _default_) o argumento recebe seus
valores de uma tábua biométrica já cadastrada no pacote. Se `TRUE`
o argumento é um vetor numérico de probabilidades
 multidecrementais obtidas com a função [calcula_multidecremento()].}
}
\value{
Um objeto da classe `lifetable` do pacote `lifecontingencies`
}
\description{
Os possíveis nomes de tábuas a serem fornecidos ao argumento `qx`
pode ser consultado com a função [mostra_tabuas_cadastradas()]
}
\examples{
tabuaAT49Masculina <- cria_tabua("AT49M")
summary(tabuaAT49Masculina)
}
