% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrai_resultados.R
\name{extrai_resultados}
\alias{extrai_resultados}
\title{Obtém os valores presentes de quantidades atuarias}
\usage{
extrai_resultados(fluxo, tipo = NULL)
}
\arguments{
\item{fluxo}{`data frame` contendo o fluxo atuarial}

\item{tipo}{Vetor de string indicando quais quantidades atuariais devem
ser retornadas. Se `NULL` (opção _default_) retorna todas as quantidades
atuariais disponíveis.}
}
\value{
Uma lista com os valores presentes das quantidades
 atuariais ou um vetor numérico unitário com o valor atual do tipo informado.
}
\description{
A função calcula o valor presente atuarial das quantidades
atuariais desejadas. Até o presente momento estão disponíveis
as seguintes quantidades: `vasf`, `vabf`, `vacf`, `vabf_bc`
`vabf_bac`, `vacf_bc`, `vacf_bac`, `rm`, `rm_bc`, `rm_bac`,
`ag` e `ra`
}
