% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Ex.R
\name{get_Ex}
\alias{get_Ex}
\title{Calcula o fator de desconto atuarial (Exn)}
\usage{
get_Ex(id, idade, sexo, invalidez, tqm, tqf, tim, tif)
}
\arguments{
\item{id}{Valor passado pela função mãe.}

\item{idade}{Valor passado pela função mãe.}

\item{sexo}{Valor passado pela função mãe.}

\item{invalidez}{Valor passado pela função mãe.}

\item{tqm}{Valor passado pela função mãe.}

\item{tqf}{Valor passado pela função mãe.}

\item{tim}{Valor passado pela função mãe.}

\item{tif}{Valor passado pela função mãe.}
}
\value{
`data frame` com os valores dos fatores de
 desconto atuarial para cada beneficiário e período
 do fluxo atuarial indicado em `id`.
}
\description{
Essa é uma função interna que tem por objetivo calcular
para cada beneficiário o desconto atuarial relativo a
cada período de tempo futuro do fluxo atuarial.
É utilizada internamente pelas funções [vabf_bc_apos()](vabf_bc_apos) e
[vacf_bc_apos()](vacf_bc_apos).
}
