% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pxT.R
\name{get_pxT}
\alias{get_pxT}
\title{Calcula probabilidades de sobrevivência multidecremental (pxT)}
\usage{
get_pxT(id, idade, sexo, tmdm, tmdf)
}
\arguments{
\item{id}{Valor passado pela função mãe.}

\item{idade}{Valor passado pela função mãe.}

\item{sexo}{Valor passado pela função mãe.}

\item{tmdm}{Valor passado pela função mãe.}

\item{tmdf}{Valor passado pela função mãe.}
}
\value{
`data frame` com os valores das probabilides
 de sobrevivência multidecrementais para cada
  beneficiário e período do fluxo atuarial indicado em `id`.
}
\description{
Esta função é usada internamente pela função [vasf()] e 
calcula para cada servidor ativo a probabilidade multidecremental
tpxT de um servidor de idade `x`, sobreviver atá a idade `x+t`
considerando um ambiente multidecremental.
}
