% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_fluxo.R
\name{id_fluxo}
\alias{id_fluxo}
\title{Cria colunas iniciais do fluxo atuarial}
\usage{
id_fluxo(dt_focal, juros)
}
\arguments{
\item{dt_focal}{Data focal da avaliação atuarial no formato 'dd/mm/yyyy'.}

\item{juros}{Taxa de juros}
}
\value{
`data frame` contendo as colunas iniciais do fluxo atuarial.
}
\description{
Cria colunas iniciais do fluxo atuarial
}
\examples{
 id_fluxo("31/12/2023", 4.93)
}
