% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importa_fluxo.R
\name{importa_fluxo}
\alias{importa_fluxo}
\title{Importa o fluxo atuarial}
\usage{
importa_fluxo(arquivo)
}
\arguments{
\item{arquivo}{Arquivo contendo o fluxo atuarial no formato `.csv`}
}
\value{
`data frame` contendo o fluxo atuarial
}
\description{
A função importa os fluxos atuariais no formato `.csv`
conforme encaminhados ao Ministério da Previdência Social - MPS
no modelo disponibilizado.
}
