% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostra_tabuas_cadastradas.R
\name{mostra_tabuas_cadastradas}
\alias{mostra_tabuas_cadastradas}
\title{Mostra nomes das tábuas cadastradas no pacote}
\usage{
mostra_tabuas_cadastradas()
}
\value{
Um `data frame` contendo os nomes das tábuas já 
 cadastradas no pacote e informações adicionais sobre as mesmas.
}
\description{
Esta função mostra os nomes das tábuas biométricas
atualmente cadastradas no pacote. Esses nomes devem
ser fornecidos às funções.
}
\examples{
mostra_tabuas_cadastradas()
}
