% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plota_fluxo.R
\name{plota_fluxo}
\alias{plota_fluxo}
\title{Elabora gráfico do fluxo atuarial da quantidade atuarial desejada.}
\usage{
plota_fluxo(fluxo, tipo = "rm")
}
\arguments{
\item{fluxo}{`data frame` contendo o fluxo atuarial}

\item{tipo}{quantidade atuarial cujo fluxo se deseja plotar. Opção _default_ `rm` (reserva matemática)}
}
\value{
Gráfico de linha com a evolução da quantidade atuarial especificada.
}
\description{
Elabora gráfico do fluxo atuarial da quantidade atuarial desejada.
}
