% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vabf_bc_apos.R
\name{vabf_bc_apos}
\alias{vabf_bc_apos}
\title{Calcula o fluxo atuarial dos benefícios futuros}
\usage{
vabf_bc_apos(
  Bx,
  inflacao = 0,
  juros,
  cb = 0,
  idade,
  invalidez = NULL,
  regra,
  sexo,
  tmg,
  tmi = invalidez,
  flx = 1:150
)
}
\arguments{
\item{Bx}{Valor do benefício pago. Vetor numérico oriundo da base cadastral.}

\item{inflacao}{Inflação de longo prazo. Usado para o calculo do fator de capacidade.
Valor numérico no formato percentual. Premissa atuarial.}

\item{juros}{axa de juros. Valor numérico em formato percentual. Premissa atuarial.}

\item{cb}{Taxa real de crescimento dos benefícios. Valor numérico em formato perncentual. Premissa atuarial.}

\item{idade}{Idade do aposentado na data focal. Vetor numérico oriundo da base cadastral.}

\item{invalidez}{Variável indicativa quanto a condição do aposentado em relação
à invalidez. Se válido ou inválido. Vetor numérico oriundo da base cadastral.}

\item{regra}{Variável indicativa da regra de aposentadoria do aposentado. Se integralidade
e paridade ou média. Vetor numérico oriundo da base cadastral.}

\item{sexo}{Sexo do aposentado. Vetor numérico oriundo da base cadastral.}

\item{tmg}{Vetor de caracteres indicando os nomes das tábuas de mortalidade a
serem utilizadas para os sexos masculino e feminino. Exemplo: `tmg = c("IBGE2022M", "IBGE2022F")`. Premissa atuarial.}

\item{tmi}{Vetor de caracteres indicando os nomes das tábuas de invalidez a
serem utilizadas para os sexos masculino e feminino. Exemplo: `tmi = "ALVARO.VINDAS"`. Premissa atuarial.}

\item{flx}{Intervalo de 1 a N para o período de projeção. Por _default_ N = 150 
para atender ao modelo de fluxo atuarial do Ministério da Previdência social - MPS.}
}
\value{
`data frame` contendo as projeções dos benefícios futuros para cada aposentado
}
\description{
A função calcula, para cada aposentado na base de dados, o fluxo futuro relativo
 às despesas com o pagamento de benefícios aos aposentados.
}
