% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vasf.R
\name{vasf}
\alias{vasf}
\title{Calcula o fluxo das projeções atuariais futuras dos salários de contribuição}
\usage{
vasf(Sx, idade, idade_prev_apos, sexo, cs, tmdm, tmdf, flx = 1:150)
}
\arguments{
\item{Sx}{Salário de contribuição na data focal da avaliação. Vetor numérico oriundo da base cadastral.}

\item{idade}{Idade dos servidores. Vetor numérico oriundo da base cadastral.}

\item{idade_prev_apos}{Idade prevista para aposentadoria. Vetor numérico oriundo da base cadastral.}

\item{sexo}{Sexo do servidor. Vetor numérico oriundo da base cadastral.}

\item{cs}{Percentual relativo à taxa real de crescimento salarial. Premissa atuarial.}

\item{tmdm}{Tábua biométrica multidecremental para o sexo masculino. Premissa atuarial.}

\item{tmdf}{Tábua biométrica multidecremental para o sexo feminino.  Premissa atuarial.}

\item{flx}{Intervalo de 1 a N para o período de projeção. Por _default_ N = 150 
para atender ao modelo de fluxo atuarial do Ministério da Previdência social - MPS.}
}
\value{
`data frame` contendo as projeções salariais para cada servidor
}
\description{
A função faz a projeção de salários futuros para uma quantidade especificada
de anos que vai do ano da avaliação até o ano anterior ao da elebibilidade ao
benefício. A projeção considera a probabilidade multidecremental.
}
